Issue 246-2: Using instead of jsps for confirmations.
