--- app/src/main/webapp/WEB-INF/jsps/editor/MemberResign.jsp
@@ -1,37 +1,0 @@
-<%--
-  Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements.  The ASF licenses this file to You
-  under the Apache License, Version 2.0 (the "License"); you may not
-  use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.  For additional information regarding
-  copyright in this work, please see the NOTICE file in the top level
-  directory of this distribution.
---%>
-<%@ include file="/WEB-INF/jsps/taglibs-struts2.jsp" %>
-
-<p>
-   <s:text name="yourWebsites.confirmResignation">
-       <s:param value="weblog" />
-   </s:text>
-</p>
-
-<table>
-<tr>
-<td>
-    <s:form action="memberResign!resign">
-		<s:hidden name="salt" />
-        <s:hidden name="weblog" />
-        <s:submit value="%{getText('generic.yes')}" />&nbsp;
-        <s:submit value="%{getText('generic.no')}" action="menu"/>
-    </s:form>
-</td>
-</tr>
-</table>
